Encode keys to str
For Redis, we want to encode keys and values to bytes. However, the
primary key also has to be part of the name in Redis, and has to be
encoded as a string. The separate primary_key_encoder in RedisTable
takes the value of the raw record, or from a query to get(), and encodes
it to a string. The default (str) works for numbers and strings.