Fetch individual records
We're going to need the ability to fetch a small number of records for
when the cache has been invalidated. For now, this requires a separate
query to be passed to the PostgresTable, where the first fetches
everything, and the second has a =ANY($1) which a list of primary keys
is expanded into.