Flush Redis before loading
Before all keys are loaded, data in Redis is now cleared completely.

This also drops the table name prefix to all Redis keys that existed
before. This means that currently, the entire namespace has to belong to
a table, and so only one table is supported. This makes deletion easier,
since there is no way to delete keys with a certain prefix.

To support multiple tables, Redis has logical databases with namespace
partitioning. There'll need to be a small storage manager that assigns
each table a logical database.