Store in sorted set
Rows in Redis are now stored in sorted sets rather than hashes. The idea
is that we always have the same keys per row, and so don't need to be
able to get specific ones. We can encode an attribute ID (which now have
fixed and known length) together with its value as single members of the
set.

This would also be possible with lists. However, sorted sets allow us to
set scores which we can then perform range queries on. This needs more
work though, since we then need to encode entire rows as elements in one
big sorted set: At the moment, keys are primary keys of rows, set
elements are attributes; we need keys to be table names, and set
elements to be rows.

On the plus side, this easily allows us to store multiple tables. But
we'll also have to do some trickery to still be able to get rows by ID.
We'll need a hash that maps each primary key of the table to its score,
so we can then do a range query on the set, then filter for exactly this
primary key.