Encode rows as set elements
Rows in Redis are now encoded into single byte strings so each one can
become one element in a sorted set for an entire table. A RedisTable now
has a name that determines the key under which this set is stored. As a
result, multiple tables are now easily supported.

A RowCodec en-/decodes records by concatenating individual attributes'
encodings, along with a 2-byte length header for the value. This imposes
a maximum length on attribute encodings.

The drawback to the new implementation is that we can no longer easily
query rows by primary key. Instead, there is a separate hash per table
that maps primary keys to the score of the row. This is queried, first,
then the sorted set for elements with that score. These are filtered
linearly for the one with the correct primary key. This implies that if
many records have the same score, looking them up will be slow.