Enable more flexible queries
We want to enable fast queries of the storage (i.e. Redis) by more than
just a primary key. On the DB (Postgres) side that's relatively easy
since we can just add filters to the query, but on the Redis side we're
constraint by having to fit this into a single float score.

To enable these queries, a custom Range class can now be provided that
specifies how to query both Postgres and Redis for a particular subset
of records (that are contiguous in some way). It contains the lower and
upper bound for the score in Redis, as well as appropriate Postgres
query parameters. A slightly-more-than trivial use case is querying a
range of timestamps represented as datetimes for Postgres, and as the
epoch timestamp in Redis.

The client has to ensure that score bounds, the DB parameters, and the
score function for the storage table match.

Marking records dirty by score is not implemented yet, so when querying
by range, stale data may be returned.

There has also been quite some refactoring and renaming of variables and
methods.