Use UTC for datetimes
Python's builtin datetimes are a bit strange without an attached
timezone. When calling timestamp() on one, the resulting number is the
epoch timestamp in the machine's local timezone (i.e. the number of
seconds since 1970-01-01 in the machine's timezone, not UTC).

This causes headaches when serializing and mixing with other datetimes,
so the responsible codec now decodes to timezone-aware UTC datetimes.
However, since asyncpg will give us naive datetimes for "timestamp
without time zone" columns, it also accepts naive datetimes when
encoding, but treats them as if they were in UTC.