Enable adjusting cache
The subset held in a cache can now be adjusted, removing old values and
adding new ones.

This required a lot of changes to the architecture and refactoring. The
Range is now called Subset and its interface is split into a basic one
for querying, and an extension that helps the cache keep track of which
values are present in cache. The record score function is now also
needed in the subset.

Each cache instance is now bound to one type of subset and takes
arguments for its instantiation instead of instances directly.