Instantiate RedisTable in CacheTable
Previously, we'd have to pass the same score function to both the
storage and cache tables (inside the subset class for the latter).
Instantiating the RedisTable from the CacheTable means we only have to
pass it once.

This unfortunately means we're locking the CacheTable into only
supporting Redis, but that was already somewhat the case due to the
specific Subset design, and it's not like there are any alternatives
anyway.

Also made the primary_key_name in all Subset classes a class member
rather than an instance member.