Change arch for multiple indexes
We want to support querying, loading, and adjusting data based on more
than one score (index). Doing that requires a bit of a major overhaul,
where what used to be the Subset is now a collection of Indexes that
enables access to DB via query and args, as well as storage via score
intervals. The key is that multiple indexes can be supported (although
at the moment it's only the mandatory primary_key index).

This is a work in progress and a few things are broken right now.