Add scratch space
The abstract storage table and the Redis implementation now offer a
scratch space, where records can be staged to be added to or deleted
from storage. Changes in scratch space are not reflected in read
operations until they are explicitly merged.

Scratch space is meant to provide a consistent view of the data even
during cache adjustments with lots of changes, all without locking read
operations for long. The Redis implementation uses a generation counter
to determine whether a record belongs to scratch space, and so just
needs to increment that counter to make the new records available.