Lock reads against writes
Inconsistent data may be returned when reading from the RedisTable while
concurrent writes are going on. To prevent this, lock reads with a
reader lock.

Unfortunately, the aiorwlock's lock classes aren't quite compatible with
asyncio.Lock, so some hackery is needed to use the writer lock in the
asyncio.Condition needed for the scratch operation exclusion.