Use scratch for cache adjusts
The cache now uses storage scratch space when it adjusts itself in order
to provide a consistent view of the storage data. This required
splitting the adjustment of the Indexes into prepare and commit phases,
which separate the index telling us what to discard and load, and us
telling the index that it's been done.