Lock conflicting operations
The 2 scratch operations in cache, adjusting and refreshing, shouldn't
step on each others toes and need to be mutexed. Similarly, we don't
want to check that a record is valid, only for it to be invalidated
before we get a chance to return it, so read operations need to be
mutexed against invalidate_record(). Lastly, invalidating a record needs
to wait for any potential refresh or adjust to finish first.