Disallow adding via invalidate_record()
Previously, it was possible to force the cache to load a new record into
storage by calling invalidate_record() on it. However, that actually has
the potential to confuse the indexes when that record is not considered
to be covered by them (i.e. covers()). The new record would be observed
(observe()), but indexes aren't required to remember observed records.
This may actually effectively leak memory, when the next call to
adjust() fails to delete it and the indexes forget about it.

So adding records like that is no longer possible. Instead, adjust() has
to be used of everything now.