Extract primary key in RecordScorer
Having the RecordScorer be able to extract the primary key from a record
(which it has to do anyway internally), we can remove knowledge of the
primary key name from the storage and cache implementations.

As a side effect, we get support for composite primary keys for free
(via a RecordScorer that just returns tuples).