Release lock for liveliness
We are yielding to the event loop in the LocalStorageTable's scratch
merge operation in order to still serve reads in between. But to
actually do that, we also have to release the writer lock. This is safe,
because after each operation, the scratch space is left in a consistent
state.