Allow disabling automatic refresh
Marking records as invalid used to always trigger a refresh on the next
read operation that includes invalid records (or a dirty index). For
tables that see lots of invalidations, that would mean many reads would
have to wait for a refresh first, making the cache pretty useless.

This adds the option to disable the automatic refresh. This means the
invalid records are still served from storage (with invalid data). The
table still stores which records are invalid, so they can be manually
refreshed.

This can be useful to increase read performance in cases where it's ok
if clients get slightly outdated data.