Observe before adjustment
Before any adjustment, a refresh is done. But refreshes used to not
observe expired and loaded records, so any records affected by the
adjustment which had also been marked as invalid wouldn't be observed.
This could lead to missing information in the indexes, e.g. with newly
added records that were marked as invalid (because they didn't exist)
and then never became visible to the indexes.

The refresh before the adjustment now observes records into the same
adjustment instance. Refreshes that are triggered otherwise don't do
this.